requirements.txt
langchain_community
langchain_google_genai
pymupdf
faiss-cpu
scikit-learn
fastapi
uvicorn